pyproject.toml
[project]
name = "Lung-and-Colon-Cancer-Classification-pytorch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "albumentations>=2.0.4",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper>=1.2.0",
    "lightning>=2.5.1.post0",
    "opencv-python>=4.11.0.86",
    "opendatasets>=0.1.22",
    "pyrootutils>=1.0.4",
    "pytorch-lightning>=2.5.1.post0",
    "torch>=2.7.0",
    "torchinfo>=1.8.0",
    "torchmetrics>=1.7.1",
    "torchvision>=0.22.0",
]

[dependency-groups]
dev = [
    "dagshub>=0.5.9",
    "hydra-optuna-sweeper>=1.2.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.6",
    "jupyter>=1.1.1",
    "mlflow>=2.22.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "rich>=14.0.0",
]
